+++ flutter_navigation_generator/example/lib/example.dart
@@ -1,5 +1,8 @@
 import 'package:example/example.navigator.dart';
+import 'package:flutter_navigation_generator_animations/flutter_navigation_generator_animations.dart';
 import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 
-@flutterNavigator
+@FlutterNavigator(
+  pageType: NativeRouteAnimation,
+)
 class MainNavigator with BaseNavigator {}