+++ flutter_navigation_generator/example/lib/main.dart
@@ -76,26 +76,32 @@ class _MyHomePageState extends State<MyHomePage> {
                 _result = await mainNavigator.goToSecondPage();
                 setState(() {});
               },
-              child: const Text("Go to page 2"),
+              child: const Text("Go to page 2 (with fade animation)"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToExampleScreenWithRequiredArgument(data: CustomModel('John', 25)),
+              onPressed: () =>
+                  mainNavigator.goToExampleScreenWithRequiredArgument(
+                      data: CustomModel('John', 25)),
               child: const Text("Go to ExampleScreenWithRequiredArgument"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.customName(id: '1', name: 'John', age: 12),
+              onPressed: () =>
+                  mainNavigator.customName(id: '1', name: 'John', age: 12),
               child: const Text("Go to RouteNameWithArguments"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.goToRouteNameWithArguments2(id: '3', name: 'Will', age: 43),
+              onPressed: () => mainNavigator.goToRouteNameWithArguments2(
+                  id: '3', name: 'Will', age: 43),
               child: const Text("Go to RouteNameWithArguments2"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.showSheetRecursiveNavigationBottomSheet(),
+              onPressed: () =>
+                  mainNavigator.showSheetRecursiveNavigationBottomSheet(),
               child: const Text("Show a bottom sheet with its own navigator"),
             ),
             ElevatedButton(
-              onPressed: () => mainNavigator.showDialogExampleDialog(text: 'hi there'),
+              onPressed: () =>
+                  mainNavigator.showDialogExampleDialog(text: 'hi there'),
               child: const Text("Show a full screen dialog"),
             ),
           ],
@@ -283,7 +289,9 @@ class RecursiveNavigationBottomSheet extends StatelessWidget {
                 child: const Text("Go to second page"),
               ),
               ElevatedButton(
-                onPressed: () => myNavigator.showSheetRecursiveNavigationBottomSheet(layers: layers + 1),
+                onPressed: () =>
+                    myNavigator.showSheetRecursiveNavigationBottomSheet(
+                        layers: layers + 1),
                 child: const Text("Open another bottom sheet"),
               ),
             ],
@@ -333,16 +341,26 @@ class ExampleScreenWithRequiredArgument extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Row(
-        children: [
-          Text(
-            data.name,
-          ),
-          Text(
-            data.age.toString(),
-          ),
-        ],
+    return Scaffold(
+      body: Center(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              data.name,
+              style: const TextStyle(fontSize: 40),
+            ),
+            Text(
+              data.age.toString(),
+              style: const TextStyle(fontSize: 40),
+            ),
+            ElevatedButton(
+              onPressed: mainNavigator.goBack,
+              child: const Text("Back"),
+            ),
+          ],
+        ),
       ),
     );
   }