+++ flutter_navigation_generator/lib/src/code_builder/import_builder.dart
@@ -22,7 +22,8 @@ class ImportBuilder {
               typeRefer(route.routeWidget, targetFile: targetFile).url,
               typeRefer(route.pageType, targetFile: targetFile).url,
               typeRefer(route.returnType, targetFile: targetFile).url,
-              ...route.parameters.map((e) => typeRefer(e, targetFile: targetFile).url),
+              ...route.parameters
+                  .map((e) => typeRefer(e, targetFile: targetFile).url),
             ])
         .expand((element) => element));
     return imports.whereType<String>().map(Directive.import);