+++ flutter_navigation_generator/lib/src/code_builder/route_builder.dart
@@ -53,7 +53,8 @@ class RouteBuilder {
     if (route.methodNameIsDefinedByAnnotation) {
       name = route.methodName;
     } else {
-      name = '$namePrefix${CaseUtil(route.methodName.replaceAll(RegExp(':[a-zA-Z0-9]*'), '')).upperCamelCase}';
+      name =
+          '$namePrefix${CaseUtil(route.methodName.replaceAll(RegExp(':[a-zA-Z0-9]*'), '')).upperCamelCase}';
     }
     return Method(
       (b) => b
@@ -67,7 +68,9 @@ class RouteBuilder {
                 ..name = p.argumentName
                 ..named = true
                 ..required = p.isRequired && route.defaultValues[p.name] == null
-                ..defaultTo = route.defaultValues[p.name] != null ? Reference(route.defaultValues[p.name]! as String).code : null
+                ..defaultTo = route.defaultValues[p.name] != null
+                    ? Reference(route.defaultValues[p.name]! as String).code
+                    : null
                 ..type = typeRefer(p),
             ),
           ),
@@ -110,34 +113,45 @@ class RouteBuilder {
         Expression path;
         final parameters = route.routeName.parametersFromRouteName;
         if (route.routeNameContainsParameters) {
-          path = Reference('RouteNames.${CaseUtil(route.routeName).camelCase}').call(
-              [],
-              (parameters.asMap().map((_, parameter) {
-                final argument = route.parameters.firstWhereOrNull((element) => element.name == parameter);
-                if (argument == null) return MapEntry(parameter, null);
-                return MapEntry(parameter, Reference(_generateParameterValue(argument)));
-              })
-                    ..removeWhere((key, value) => value == null))
-                  .cast());
+          path = Reference('RouteNames.${CaseUtil(route.routeName).camelCase}')
+              .call(
+                  [],
+                  (parameters.asMap().map((_, parameter) {
+                    final argument = route.parameters.firstWhereOrNull(
+                        (element) => element.name == parameter);
+                    if (argument == null) return MapEntry(parameter, null);
+                    return MapEntry(parameter,
+                        Reference(_generateParameterValue(argument)));
+                  })
+                        ..removeWhere((key, value) => value == null))
+                      .cast());
         } else {
           path = Reference('RouteNames.${CaseUtil(route.routeName).camelCase}');
         }
-        final queryParameters =
-            route.parameters.where((element) => element.className != 'Key' && !parameters.contains(element.argumentName)).toList().asMap().map((_, p) => MapEntry(
+        final queryParameters = route.parameters
+            .where((element) =>
+                element.className != 'Key' &&
+                !parameters.contains(element.argumentName))
+            .toList()
+            .asMap()
+            .map((_, p) => MapEntry(
                   "'${p.argumentName}'",
                   _generateParameterValue(p),
                 ));
 
         final bodyCall = TypeReference(
           (b) => b
-            ..symbol = 'navigatorKey.currentState?.${route.navigationType.navigatorMethod}'
+            ..symbol =
+                'navigatorKey.currentState?.${route.navigationType.navigatorMethod}'
             ..types.addAll(
               [
                 const Reference('dynamic'),
                 if (route.navigationType == NavigationType.pushReplacement ||
                     route.navigationType == NavigationType.popAndPush ||
-                    route.navigationType == NavigationType.restorablePushReplacement ||
-                    route.navigationType == NavigationType.restorablePopAndPush) ...[
+                    route.navigationType ==
+                        NavigationType.restorablePushReplacement ||
+                    route.navigationType ==
+                        NavigationType.restorablePopAndPush) ...[
                   const Reference('dynamic'),
                 ],
               ],
@@ -156,19 +170,28 @@ class RouteBuilder {
                       'queryParameters': Reference('$queryParameters'),
                     },
                   ).property('toString()'),
-            if (route.navigationType == NavigationType.pushAndReplaceAll || route.navigationType == NavigationType.restorablePushAndReplaceAll) ...[
+            if (route.navigationType == NavigationType.pushAndReplaceAll ||
+                route.navigationType ==
+                    NavigationType.restorablePushAndReplaceAll) ...[
               const Reference('(_) => false'),
             ],
           ],
           {
-            'arguments': Reference('${route.parameters.asMap().map((_, p) => MapEntry("'${p.argumentName}'", p.argumentName))}'),
+            'arguments': Reference(
+                '${route.parameters.asMap().map((_, p) => MapEntry("'${p.argumentName}'", p.argumentName))}'),
           },
         );
         Code body;
         if (route.returnType != null) {
           body = Block((b) => b
-            ..statements.add(declareFinal('result', type: const Reference('dynamic')).assign(bodyCall.awaited).statement)
-            ..statements.add(const Reference('result').asA(typeRefer(route.returnType, forceNullable: true)).returned.statement));
+            ..statements.add(
+                declareFinal('result', type: const Reference('dynamic'))
+                    .assign(bodyCall.awaited)
+                    .statement)
+            ..statements.add(const Reference('result')
+                .asA(typeRefer(route.returnType, forceNullable: true))
+                .returned
+                .statement));
         } else {
           body = bodyCall.code;
         }
@@ -194,16 +217,22 @@ class RouteBuilder {
       body: TypeReference(
         (b) => b
           ..symbol = bodyCall
-          ..types.add(route.returnType == null ? const Reference('dynamic') : typeRefer(route.returnType!)),
+          ..types.add(route.returnType == null
+              ? const Reference('dynamic')
+              : typeRefer(route.returnType!)),
       ).call(
         [],
         {
           'widget': Reference(
-            route.constructorName == route.routeWidget.className || route.constructorName.isEmpty
+            route.constructorName == route.routeWidget.className ||
+                    route.constructorName.isEmpty
                 ? route.routeWidget.className
                 : '${route.routeWidget.className}.${route.constructorName}',
             typeRefer(route.routeWidget).url,
-          ).call([], route.parameters.asMap().map((_, p) => MapEntry(p.argumentName, Reference(p.argumentName)))),
+          ).call(
+              [],
+              route.parameters.asMap().map((_, p) =>
+                  MapEntry(p.argumentName, Reference(p.argumentName)))),
         },
       ).code,
     );
@@ -231,7 +260,8 @@ class RouteBuilder {
             ..name = 'goBack'
             ..lambda = true
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.pop').call([]).code,
+            ..body =
+                const Reference('navigatorKey.currentState?.pop').call([]).code,
         ),
         Method(
           (b) => b
@@ -245,7 +275,8 @@ class RouteBuilder {
                 ..type = const Reference('T?'),
             ))
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.pop').call([const Reference('result')]).code,
+            ..body = const Reference('navigatorKey.currentState?.pop')
+                .call([const Reference('result')]).code,
         ),
         Method(
           (b) => b
@@ -259,7 +290,8 @@ class RouteBuilder {
                   ..requiredParameters.add(const Reference('Route<dynamic>'))),
             ))
             ..returns = const Reference('void')
-            ..body = const Reference('navigatorKey.currentState?.popUntil').call([const Reference('predicate')]).code,
+            ..body = const Reference('navigatorKey.currentState?.popUntil')
+                .call([const Reference('predicate')]).code,
         ),
         Method(
           (b) => b
@@ -288,7 +320,8 @@ class RouteBuilder {
             ..returns = const Reference('Future<T?>')
             ..body = const Reference('showDialog<T>').call([], {
               'context': const Reference('navigatorKey.currentContext!'),
-              'builder': const Reference('(_) => widget ?? const SizedBox.shrink()'),
+              'builder':
+                  const Reference('(_) => widget ?? const SizedBox.shrink()'),
             }).code,
         ),
         Method(
@@ -306,7 +339,8 @@ class RouteBuilder {
             ..returns = const Reference('Future<T?>')
             ..body = const Reference('showModalBottomSheet<T>').call([], {
               'context': const Reference('navigatorKey.currentContext!'),
-              'builder': const Reference('(_) => widget ?? const SizedBox.shrink()'),
+              'builder':
+                  const Reference('(_) => widget ?? const SizedBox.shrink()'),
             }).code,
         ),
       ];