+++ flutter_navigation_generator/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for web parameters as query parameters and url path parameters
 - Added support for methodName
+- Added a package (flutter_navigation_generator_animations) with NativeRouteAnimation to animate between pages in a native way (Android scales, iOS uses CupertinoPageTransition, web doesn't animate, other platforms use FadeTransition)
 
 ## 1.0.4+1 (2024-03-06)